pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "skilllint"
version = "0.1.0"
description = "Security scanner for agent skills"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "xujiajun-gmail" }]
keywords = ["security", "agent", "skills", "scanner", "cli", "llm"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
  "typer>=0.16.0",
  "rich>=14.0.0",
  "pydantic>=2.11.0",
  "pyyaml>=6.0.2",
  "httpx>=0.28.0",
  "jinja2>=3.1.6",
]

[project.optional-dependencies]
llm = ["openai>=1.76.0"]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.11.0",
  "mypy>=1.15.0",
]

[project.scripts]
skilllint = "skilllint.cli:app"

[tool.hatch.build]
include = ["src/skilllint/**/*.yaml"]

[tool.hatch.build.targets.wheel]
packages = ["src/skilllint"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008", "E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
